serverless-backend/requirements.txt
azure-functions
azure-data-tables
google-genai